My AuthAuth solution

+++ html/admin.js
@@ -8,17 +8,30 @@ const updateUsers = () => {
       row.innerHTML = `<td>${u}</td>
                        <td><input type="checkbox"${a ? ' checked' : ''} /></td>
                        <td><button>Delete</button></td>`
-      row.querySelector('input').addEventListener('change', console.log)
+      row.querySelector('input').addEventListener('change', toggleAdmin(u, a))
       row.querySelector('button').addEventListener('click', deleteUser(u))
       tbody.appendChild(row)
     }))
+    .catch(console.log)
 }
 updateUsers()
 
-const deleteUser = (u) => () => fetch(
-    '/api/admin?user=' + u,
-  { method: 'DELETE' }
-).then(updateUsers)
+const deleteUser = (u) => () => {
+  if(confirm('Really delete ' + u + '?')) {
+    fetch(
+        '/api/admin?user=' + u,
+      { method: 'DELETE' }
+    ).then(updateUsers)
+     .catch(console.log)
+  }
+}
+const toggleAdmin = (u, a) => () => {
+  fetch('/api/admin', {
+    method: 'PUT',
+    body: JSON.stringify({ user: u, admin: !a })
+  }).then(updateUsers)
+    .catch(console.log)
+}
 
 document.getElementById('adduser').addEventListener(
   'click',
@@ -29,5 +42,11 @@ document.getElementById('adduser').addEventListener(
       method: 'POST',
       body: JSON.stringify({ user, pass })
     }).then(updateUsers)
+      .catch(console.log)
   }
+)
+
+document.getElementById('logout').addEventListener(
+  'click',
+  () => fetch('/logout')
 )
